Fixes #9417: Correct test on expected reports and clean code
